29.05 state
adapters added

+++ app/src/main/java/com/example/nework/dto/Post.kt
@@ -79,9 +79,9 @@ enum class AttachmentType {
 
 data class Coords(
     @SerializedName("lat")
-    val lat: Int,
+    val lat: Double,
     @SerializedName("long")
-    val long: Int
+    val long: Double
 )
 
 data class Ad(
